test: maths tests refactoring
* moving strategies to utils so that they are more general purpose than stateful testing.

* updated `test_get_y.py` to use more hypothesis features and minor refactors.

* removed duplicate code from `test_newton_D.py`
