+++ tests/unitary/math/test_get_y.py
@@ -1,32 +1,21 @@
-# flake8: noqa
-import time
 from decimal import Decimal
 
 import boa
 import pytest
 from hypothesis import event, given, note, settings
-from hypothesis import strategies as st
+from hypothesis.strategies import integers
 
-from tests.utils.constants import MAX_GAMMA, MIN_GAMMA
+from tests.utils.strategies import A, gamma
 
-N_COINS = 2
-# MAX_SAMPLES = 1000000  # Increase for fuzzing
+# you might want to increase this when fuzzing locally
 MAX_SAMPLES = 10000
 N_CASES = 32
 
-A_MUL = 10000
-MIN_A = int(N_COINS**N_COINS * A_MUL / 10)
-MAX_A = int(N_COINS**N_COINS * A_MUL * 1000)
-
-
-pytest.current_case_id = 0
-pytest.negative_sqrt_arg = 0
-pytest.gas_original = 0
-pytest.gas_new = 0
-pytest.t_start = time.time()
-
 
 def inv_target_decimal_n2(A, gamma, x, D):
+    """Computes the inavriant (F) as described
+    in the whitepaper.
+    """
     N = len(x)
 
     x_prod = Decimal(1)
@@ -56,10 +45,10 @@ def test_get_y_revert(math_contract):
     D = 224824250915890636214130540882688
     i = 0
 
-    with boa.reverts():
+    with boa.reverts(dev="unsafe values A"):
         math_contract.newton_y(a, gamma, x, D, i)
 
-    with boa.reverts():
+    with boa.reverts(dev="unsafe values A"):
         math_contract.get_y(a, gamma, x, D, i)
 
 
@@ -67,22 +56,22 @@ def test_get_y_revert(math_contract):
     "_tmp", range(N_CASES)
 )  # Parallelisation hack (more details in folder's README)
 @given(
-    A=st.integers(min_value=MIN_A, max_value=MAX_A),
-    D=st.integers(
+    A=A,
+    gamma=gamma,
+    D=integers(
         min_value=10**18, max_value=10**14 * 10**18
     ),  # 1 USD to 100T USD
-    xD=st.integers(
+    xD=integers(
         min_value=10**17 // 2, max_value=10**19 // 2
-    ),  # <- ratio 1e18 * x/D, typically 1e18 * 1
-    yD=st.integers(
+    ),  # ratio 1e18 * x/D, typically 1e18 * 1
+    yD=integers(
         min_value=10**17 // 2, max_value=10**19 // 2
-    ),  # <- ratio 1e18 * y/D, typically 1e18 * 1
-    gamma=st.integers(min_value=MIN_GAMMA, max_value=MAX_GAMMA),
-    j=st.integers(min_value=0, max_value=1),
+    ),  # ratio 1e18 * y/D, typically 1e18 * 1
+    j=integers(min_value=0, max_value=1),
 )
 @settings(max_examples=MAX_SAMPLES, deadline=None)
 def test_get_y(math_unoptimized, math_optimized, A, D, xD, yD, gamma, j, _tmp):
-    pytest.current_case_id += 1
+    # pytest.current_case_id += 1
     X = [D * xD // 10**18, D * yD // 10**18]
 
     A_dec = Decimal(A) / 10000 / 4
@@ -95,20 +84,26 @@ def test_get_y(math_unoptimized, math_optimized, A, D, xD, yD, gamma, j, _tmp):
     try:
         result_original = math_unoptimized.newton_y(A, gamma, X, D, j)
     except Exception as e:
+        event("hit unsafe for unoptimizied")
         if "unsafe value" in str(e):
-            assert not "gamma" in str(e)
+            assert "gamma" not in str(e)
             assert gamma > 2 * 10**16
             return
         else:  # Did not converge?
             raise
-    pytest.gas_original += math_unoptimized._computation.get_gas_used()
+    unoptimized_gas = math_unoptimized._computation.net_gas_used
+    event(
+        "unoptimizied implementation used {:.0e} gas".format(unoptimized_gas)
+    )
 
     try:
         result_get_y, K0 = math_optimized.get_y(A, gamma, X, D, j)
     except Exception as e:
+        event("hit unsafe for optimizied")
         if "unsafe value" in str(e):
-            # The only possibility for old one to not revert and new one to revert is to have
-            # very small difference near the unsafe y value boundary.
+            # The only possibility for old one to not revert and
+            # new one to revert is to have very small difference
+            # near the unsafe y value boundary.
             # So, here we check if it was indeed small
             lim_mul = 100 * 10**18
             if gamma > 2 * 10**16:
@@ -123,7 +118,8 @@ def test_get_y(math_unoptimized, math_optimized, A, D, xD, yD, gamma, j, _tmp):
                 raise
         else:
             raise
-    pytest.gas_new += math_optimized._computation.get_gas_used()
+    optimized_gas = math_optimized._computation.net_gas_used
+    event("optimizied implementation used {:.0e} gas".format(optimized_gas))
 
     note(
         "{"
@@ -133,14 +129,7 @@ def test_get_y(math_unoptimized, math_optimized, A, D, xD, yD, gamma, j, _tmp):
 
     if K0 == 0:
         event("fallback to newton_y")
-        pytest.negative_sqrt_arg += 1
         return
-
-    if pytest.current_case_id % 1000 == 0:
-        print(
-            f"--- {pytest.current_case_id}\nPositive dy frac: {100*pytest.negative_sqrt_arg/pytest.current_case_id:.1f}%\t{time.time() - pytest.t_start:.1f} seconds.\n"
-            f"Gas advantage per call: {pytest.gas_original//pytest.current_case_id} {pytest.gas_new//pytest.current_case_id}\n"
-        )
 
     assert abs(result_original - result_get_y) <= max(
         10**4, result_original / 1e8
