+++ tests/unitary/pool/stateful/stateful_base.py
@@ -10,12 +10,12 @@ from hypothesis.stateful import (
     rule,
 )
 from hypothesis.strategies import integers
-from strategies import address
-from strategies import pool as pool_strategy
 
 from contracts.main import CurveTwocryptoFactory as factory
 from contracts.mocks import ERC20Mock as ERC20
 from tests.utils.constants import UNIX_DAY
+from tests.utils.strategies import address
+from tests.utils.strategies import pool as pool_strategy
 from tests.utils.tokens import mint_for_testing
